+++ modules/hanna-css/src/css/ContentArticle.css.ts
@@ -16,7 +16,27 @@ export default css`
   @media screen {
     .ContentArticle {
       ${SeenEffect__fadeup}
-      ${centerColumnsStyling}
+      ${centerColumnsStyling({ allowed: ['.ContentArticle__header'] })}
+    }
+
+    .ContentArticle__header {
+      display: flex;
+      flex-flow: column;
+    }
+    .ContentArticle__header > .PageHeading,
+    .ContentArticle__header > .Heading {
+      order: 1;
+    }
+    .ContentArticle__header > .Heading:not(:last-child) {
+      margin-top: 0;
+    }
+
+    // Support for simpler (and legacy) usage patterns where __heaeder
+    // is NOT present and .Heading is preceeded by .ArticlaMeta, etc.
+    .ContentArticle > .Heading:nth-child(2),
+    .ContentArticle > .Heading:nth-child(3),
+    .ContentArticle > .Heading:nth-child(4) {
+      margin-top: 0;
     }
   }
 `;